requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
docopt==0.6.2
Flask==1.1.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
pipreqs==0.4.10
protobuf==3.6.1
public==2019.4.13
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
WTForms==2.2.1
yarg==0.1.9
